refactor(configparse): rename to configfile
