+++ internal/configfile/json.go
@@ -1,4 +1,4 @@
-package configparse
+package configfile
 
 import (
 	"github.com/benchttp/engine/runner"
