+++ internal/configfile/parse.go
@@ -1,4 +1,4 @@
-package configparse
+package configfile
 
 import (
 	"errors"
