+++ internal/configfile/parser_internal_test.go
@@ -1,4 +1,4 @@
-package configparse
+package configfile
 
 import (
 	"errors"
