+++ internal/configfile/testdata/extends/extends-circular-0.yml
