+++ internal/configfile/testdata/extends/extends-valid-child.yml
