+++ internal/configfile/testdata/extends/extends-valid-parent.yml
